requirements.txt
bitcoin==1.1.42
docopt==0.6.2
ecdsa==0.13
protobuf==3.5.2.post1
pycrypto==2.6.1
pykwalify==1.6.1
PySocks==1.6.8
python-dateutil==2.7.3
PyYAML==3.12
six==1.11.0
